--- a/ai-engine/.env.example
+++ b/ai-engine/.env.example
@@ -63,3 +63,8 @@ BURST_LIMIT=10
 # Testing
 MOCK_AI_RESPONSES=false
 TEST_DATA_DIR=./test_data
+
+# Ollama Configuration
+USE_OLLAMA=false
+OLLAMA_MODEL=llama3.2
+OLLAMA_BASE_URL=http://localhost:11434