--- a/ai-engine/pytest.ini
+++ b/ai-engine/pytest.ini
@@ -6,7 +6,7 @@ python_functions = test_*
 asyncio_mode = auto
 asyncio_default_fixture_loop_scope = function
 asyncio_default_test_loop_scope = function
-addopts = 
+addopts =
     --strict-markers
     --strict-config
     --verbose
@@ -17,6 +17,7 @@ addopts =
     --cov-report=html:htmlcov
     --cov-report=xml
     --cov-fail-under=40
+    --timeout=300
 markers =
     unit: Unit tests
     integration: Integration tests