--- a/ai-engine/src/agents/packaging_agent.py
+++ b/ai-engine/src/agents/packaging_agent.py
@@ -2,7 +2,7 @@
 Packaging Agent for assembling converted components into .mcaddon packages
 """
 
-from typing import Dict, List, Optional
+from typing import Dict, List, Optional, Any
 
 import logging
 import json