… .github/copilot-instructions.md
… .github/security-config-guide.md
… .github/self-hosted-runner-setup.md
… .github/workflows/build-base-images.yml
… .github/workflows/ci-optimized.yml
… .github/workflows/ci.yml
… .github/workflows/test-optimization.yml
… CI_OPTIMIZATION_README.md
… CLAUDE.md
… ai-engine/requirements-dev.txt
… ai-engine/requirements.txt
… ai-engine/src/agents/asset_converter.py
… ai-engine/src/agents/logic_translator.py
… ai-engine/src/agents/qa_agent.py
… ai-engine/src/agents/qa_validator.py
… ai-engine/src/agents/rag_agents.py
… ai-engine/src/agents/validation_agent.py
… ai-engine/src/benchmarking/metrics_collector.py
… ai-engine/src/benchmarking/performance_system.py
… ai-engine/src/crew/conversion_crew.py
… ai-engine/src/engines/comparison_engine.py
… ai-engine/src/main.py
… ai-engine/src/testing/behavioral_framework.py
… ai-engine/src/training_manager.py
… ai-engine/src/utils/rate_limiter.py
… ai-engine/src/utils/vector_db_client.py
… ai-engine/test_structure.py
… ai-engine/tests/conftest.py
… ai-engine/tests/crew/test_rag_crew.py
… ai-engine/tests/fixtures/conversion_test_mods/mowzies_mobs/mowzies_mobs.jar